ButtonOrLink: barrel structure
