+++ client/common/ButtonOrLink/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ButtonOrLink';
+export type { ButtonOrLinkProps, Ref } from './ButtonOrLink.types';
